# specs/openapi.yaml
openapi: 3.1.0
info:
  title: VirtualRadio API
  version: 1.0.0
  description: |
    API REST de **VirtualRadio**: plataforma que genera de forma automatizada
    programas de radio satíricos (estilo WCTR) para videojuegos de simulación
    (Farming Simulator 25, Euro Truck Simulator 2, etc.).

    El backend orquesta una biblioteca musical local, un universo narrativo
    compartido (noticias, comerciales, personajes y memoria narrativa) y un
    pipeline de generación por agentes que produce episodios de audio listos
    para reproducir.

    ## Autenticación
    Autenticación mediante **JWT** (Flask-JWT-Extended). `POST /auth/login`
    emite un *access token* y un *refresh token*; `POST /auth/refresh` renueva
    el *access token* a partir de un *refresh token* válido. Todos los endpoints
    salvo los de `auth` y `health` requieren `Authorization: Bearer <access_token>`.

    ## Autorización (RBAC)
    Modelo deliberadamente simple: un único rol de sistema `USER` y un único
    scope efectivo `own`. Cada consulta se filtra por `owner_id = current_user.id`.
    El `owner_id` se toma **siempre del token**, nunca del body. El acceso a un
    recurso de otro usuario devuelve **404** (no 403) para no revelar su existencia.

    ## Contrato de respuestas
    - Éxito: `{ "data": ..., "meta": {...} }`
    - Error: `{ "error": { "code", "message", "details" } }`
  contact:
    name: Equipo de Plataforma / Backend
  license:
    name: Propietario
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /stations:
    get:
      tags: [Stations]
      summary: Listar estaciones
      description: 'Permiso: `station:read:own`.'
      operationId: listStations
      responses:
        '200':
          description: Lista de estaciones del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Stations]
      summary: Crear estación
      description: 'Permiso: `station:create:own`. El `owner_id` se toma del token.'
      operationId: createStation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationInput'
      responses:
        '201':
          description: Estación creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /stations/suggest:
    post:
      tags: [Stations]
      summary: Sugerir estación por IA
      description: |
        Genera una propuesta de estación con el LLM (con fallback procedural
        determinista si el LLM no responde). Permiso: `station:suggest:own`.
      operationId: suggestStation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestRequest'
      responses:
        '200':
          description: Sugerencia de estación (no persistida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /stations/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Stations]
      summary: Obtener estación
      description: 'Permiso: `station:read:own`.'
      operationId: getStation
      responses:
        '200':
          description: Estación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Stations]
      summary: Actualizar estación
      description: 'Permiso: `station:update:own`.'
      operationId: updateStation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationInput'
      responses:
        '200':
          description: Estación actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Stations]
      summary: Eliminar estación
      description: 'Permiso: `station:delete:own`. Borra en cascada episodios y jobs asociados.'
      operationId: deleteStation
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Episodes
  # ---------------------------------------------------------------------------
...
  /story-events:
    get:
      tags: [Story Events]
      summary: Listar eventos narrativos
      description: 'Permiso: `story_event:read:own`.'
      operationId: listStoryEvents
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StoryStatus'
      responses:
        '200':
          description: Lista de eventos narrativos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryEventList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Story Events]
      summary: Crear evento narrativo
      description: 'Permiso: `story_event:create:own`.'
      operationId: createStoryEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoryEventInput'
      responses:
        '201':
          description: Evento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryEventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /story-events/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Story Events]
      summary: Obtener evento narrativo
      description: 'Permiso: `story_event:read:own`.'
      operationId: getStoryEvent
      responses:
        '200':
          description: Evento narrativo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryEventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Story Events]
      summary: Actualizar evento narrativo
      description: 'Permiso: `story_event:update:own`.'
      operationId: updateStoryEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoryEventInput'
      responses:
        '200':
          description: Evento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryEventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Story Events]
      summary: Eliminar evento narrativo
      description: 'Permiso: `story_event:delete:own`.'
      operationId: deleteStoryEvent
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Universe
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token JWT emitido por `/auth/login` o `/auth/refresh`.'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Identificador UUID del recurso
      schema:
        type: string
        format: uuid
  responses:
    NoContent:
      description: Operación realizada con éxito (sin contenido)
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Token inválido o ausente
              details: {}
    NotFound:
      description: |
        El recurso no existe o pertenece a otro usuario (se devuelve 404 en
        ambos casos para no revelar su existencia).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Recurso no encontrado
              details: {}
    ValidationError:
      description: Error de validación de los datos enviados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Los datos enviados no son válidos
              details:
                name: ["El campo es obligatorio"]
  schemas:
...
    Meta:
      type: object
      description: Metadatos de la respuesta (paginación, conteos, etc.)
      additionalProperties: true
      example: {}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Código de error legible por máquina
              example: VALIDATION_ERROR
            message:
              type: string
              description: Descripción del error
            details:
              type: object
              description: Detalles adicionales (ej. errores por campo)
              additionalProperties: true
    Timestamps:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true

    # ---- Enums ----------------------------------------------------------------
...
    StoryStatus:
      type: string
      description: Estado de un arco narrativo
      enum:
        - active
        - resolved
...
    SuggestRequest:
      type: object
      description: Pista/contexto opcional para guiar la sugerencia del LLM.
      properties:
        prompt:
          type: string
          description: Indicación libre para orientar la generación
        context:
          type: object
          description: Contexto adicional (ej. estación, tema)
          additionalProperties: true

    # ---- Stations -------------------------------------------------------------
    StationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 120
          description: Nombre de la estación (único por usuario)
        host_name:
          type: string
          maxLength: 120
          nullable: true
        description:
          type: string
          nullable: true
        personality:
          type: string
          nullable: true
        frequency:
          type: string
          maxLength: 20
          nullable: true
          example: "99.1 FM"
        emoji:
          type: string
          maxLength: 16
          nullable: true
        color:
          type: string
          maxLength: 9
          nullable: true
          example: "#FFAA00"
        intro_templates:
          type: array
          items:
            type: string
          default: []
        outro_templates:
          type: array
          items:
            type: string
          default: []
    Station:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            owner_id:
              type: string
              format: uuid
              readOnly: true
        - $ref: '#/components/schemas/StationInput'
        - $ref: '#/components/schemas/Timestamps'
    StationResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Station'
        meta:
          $ref: '#/components/schemas/Meta'
    StationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Station'
        meta:
          $ref: '#/components/schemas/Meta'

    # ---- Episodes -------------------------------------------------------------
...
    StoryEventInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          nullable: true
        related_characters:
          type: string
          nullable: true
          description: Personajes implicados (lista)
        status:
          $ref: '#/components/schemas/StoryStatus'
        resolved_at:
          type: string
          format: date-time
          nullable: true
    StoryEvent:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            owner_id:
              type: string
              format: uuid
              readOnly: true
        - $ref: '#/components/schemas/StoryEventInput'
        - $ref: '#/components/schemas/Timestamps'
    StoryEventResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/StoryEvent'
        meta:
          $ref: '#/components/schemas/Meta'
    StoryEventList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/StoryEvent'
        meta:
          $ref: '#/components/schemas/Meta'

    # ---- Universe -------------------------------------------------------------
...